requirements.txt
sqlalchemy>=2.0.29
psycopg[binary]>=3.2.0
python-chess>=1.999
pydantic-settings>=2.8.0
tqdm>=4.66.0
